create_Main_redux
