--- src/redux/AddWater/selectors.js
@@ -1,2 +1,0 @@
-export const selectLoading = (state) => state.water.isLoading;
-export const selectWaterList = (state) => state.water.waterList;
