+++ src/redux/monthWater/monthWaterThunk.js
@@ -1,6 +1,5 @@
-import { createAsyncThunk } from '@reduxjs/toolkit';
-import { axiosLoader } from '../../axiosConfig/axiosLoader';
-
+import { createAsyncThunk } from "@reduxjs/toolkit";
+import { axiosLoader } from "../../axiosConfig/axiosLoader";
 
 // запит працює повертає обєкт з обєктами даних
 // {
@@ -16,10 +15,12 @@ import { axiosLoader } from '../../axiosConfig/axiosLoader';
 //   }
 // }
 export const getMonthWater = createAsyncThunk(
-  'monthWater/getMonthWater',
+  "monthWater/getMonthWater",
   async (selectDate, thunkAPI) => {
     try {
-      const { data } = await axiosLoader.get(`/water/month?month=${selectDate}`);
+      const { data } = await axiosLoader.get(
+        `/water/month?month=${selectDate}`
+      );
       return data;
     } catch (error) {
       return thunkAPI.rejectWithValue(error.message);
